requirements.txt
arrow==0.15.2
binaryornot==0.4.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
Django==2.2.5
django-crispy-forms==1.8.1
future==0.17.1
gitdb2==2.0.6
GitPython==3.0.2
idna==2.8
Jinja2==2.10.1
jinja2-time==0.2.0
MarkupSafe==1.1.1
poyo==0.5.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.3.0
urllib3==1.25.6
whichcraft==0.6.1
